refactor : apikey에 대한 applicationService와 DomainService를 나누도록 변경
